Refactor into separate directory
