+++ noise/codec.go
@@ -1,0 +1,7 @@
+package noise
+
+import (
+	amino "github.com/tendermint/go-amino"
+)
+
+var cdc = amino.NewCodec()
